requirements.txt
streamlit
tabula-py
openpyxl
Pillow
